requirements/requirements-dev.txt
-r requirements.txt

ipython==7.25.0
django-extensions==3.1.1
Werkzeug==2.0.1
